schema-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Schema generation
openapi-schema-pydantic==1.2.4
pydantic-settings==2.2.1
ruamel.yaml==0.18.16
